rpc-store: Rename the schema keys module to primitives
The `schema::keys` module holds encodable building blocks reused
across many CFs -- `U64Be`, `U64Varint`, `UnitKey`, `write_uleb128`,
and `read_struct_tag` -- which are key *and* value primitives rather
than keys specifically. Rename the module (and its file) to
`primitives` to reflect that, and update every reference.

This is a pure rename: the module's contents are unchanged.
