+++ crates/sui-rpc-store/src/indexer/live_objects.rs
@@ -31,8 +31,8 @@ use crate::indexer::Schema;
 use crate::indexer::Store;
 use crate::indexer::checkpoint_input_objects;
 use crate::indexer::checkpoint_output_objects;
-use crate::schema::keys::U64Varint;
 use crate::schema::live_objects;
+use crate::schema::primitives::U64Varint;
 
 /// Pipeline marker for `live_objects`.
 pub struct LiveObjects;
