+++ crates/sui-rpc-store/src/indexer/object_by_type.rs
@@ -21,8 +21,8 @@ use crate::indexer::Schema;
 use crate::indexer::Store;
 use crate::indexer::checkpoint_input_objects;
 use crate::indexer::checkpoint_output_objects;
-use crate::schema::keys::U64Varint;
 use crate::schema::object_by_type;
+use crate::schema::primitives::U64Varint;
 
 /// Pipeline marker for `object_by_type`.
 pub struct ObjectByType;
