+++ crates/sui-rpc-store/src/indexer/transactions.rs
@@ -15,7 +15,7 @@ use sui_types::full_checkpoint_content::Checkpoint;
 use crate::indexer::Schema;
 use crate::indexer::Store;
 use crate::indexer::tx_seq_at;
-use crate::schema::keys::U64Be;
+use crate::schema::primitives::U64Be;
 use crate::schema::transactions;
 
 /// Pipeline marker for `transactions`.
