+++ crates/sui-rpc-store/src/schema/checkpoint_contents.rs
@@ -14,7 +14,7 @@ use sui_types::messages_checkpoint::CheckpointContents;
 use sui_types::messages_checkpoint::CheckpointSequenceNumber;
 
 use crate::proto::StoredCheckpointContents;
-use crate::schema::keys::U64Be;
+use crate::schema::primitives::U64Be;
 
 pub const NAME: &str = "checkpoint_contents";
