+++ crates/sui-rpc-store/src/schema/effects.rs
@@ -11,7 +11,7 @@ use sui_types::effects::TransactionEffects;
 use sui_types::storage::ObjectKey;
 
 use crate::proto::StoredEffects;
-use crate::schema::keys::U64Be;
+use crate::schema::primitives::U64Be;
 
 pub const NAME: &str = "effects";
